Reorganize term module and submodules
